fix: cast grid coords to float32 for precision in grid_sampler_3d
Root cause: grid coordinates were loaded as float16, causing
significant precision loss in (gx+1)*(size-1)*0.5 computation
(~50 range * float16 epsilon). This cascaded through the 7-lerp
trilinear interpolation chain, producing errors up to 0.019 vs
float64 reference.

Fix: explicitly cast gx, gy, gz to float32 after load, matching
PyTorch's internal behavior (always uses float32 for coordinate
computation regardless of input dtype).

Also adjust float16 atol from 1e-2 down to 3e-3 — with float32
coordinates, the dominant error is only float16 corner precision.

+++ src/flag_gems/ops/grid_sampler_3d.py
@@ -81,9 +81,9 @@ def grid_sampler_3d_kernel(
         + oh_idx * grid_strides_oh
         + ow_idx * grid_strides_ow
     )
-    gx = tl.load(grid_ptr + grid_offset + 0 * grid_stride_xyz)
-    gy = tl.load(grid_ptr + grid_offset + 1 * grid_stride_xyz)
-    gz = tl.load(grid_ptr + grid_offset + 2 * grid_stride_xyz)
+    gx = tl.load(grid_ptr + grid_offset + 0 * grid_stride_xyz).to(tl.float32)
+    gy = tl.load(grid_ptr + grid_offset + 1 * grid_stride_xyz).to(tl.float32)
+    gz = tl.load(grid_ptr + grid_offset + 2 * grid_stride_xyz).to(tl.float32)
 
     # Convert grid coordinates from [-1, 1] to input space
     if align_corners:
